Make top-level errors gracefully exit
It turns out that at present, the behavior of what happens when a `main`
function that returns an error `Result` value is not nailed down,
and in some environments, this results in a panic and hard crash!

So instead of relying on that unspecified behavior, in `main`,
explicitly check the final value and call `std::process::exit` in case
of error.

Also refine formatting in several errors that include path information:
instead of using the `Debug` trait, use the `.display()` method.

Fixes #9.