Build a multiplatform Docker image for release tags
This new workflow builds a `noseyparker` image that supports x86_64 and ARM64.
The image is built using QEMU emulation under the hood.

Fixes #39.  Mostly, anyway: the multiplatform image is very slow to
build (about 3 hours), and so for now it is only being done on release
tags or when manually requested, not every single time.
Additionally, the only tag being applied right now is `experimental-multiplatform`.