Use GitHub Actions to build a multiplatform Docker image (#45)
This commit adds a new `Release Artifacts` workflow in GitHub Actions that runs when manually requested and when a new `v*.*.*` tag is created on `main`. The new workflow creates a multiplatform Docker image that supports x86_64 and ARM64.

Note that building the Docker image is very slow, taking about 3 hours on a hosted runner. This is due mainly to the use of QEMU to build the image slice for the non-native architecture (like ARM64 from an x86_64 machine). Because of this slowness, the workflow is only run for tags, and not on every commit.

Fixes #39.