Handle missing blob metadata gracefully.
The version 2 datastore schema was recently introduced.
This added a couple new tables for recording blob metadata.

When migrating an old version 1 datastore to the version 2 schema,
it is not possible to populate the new tables with data for blobs from
old matches. So in that case, blob metadata would be missing.

Prior to this commit, the `report` command would fail to show any match
details when run on a datastore that was migrated from the earlier
version 1 datastore schema. This fixes that, causing `report` to still
function with graceful degradation when blob metadata is missing.