Collect additional metadata from scanned Git repositories (#66)
The `scan` command now collects additional metadata about blobs found within Git repositories. Specifically, for each blob found in Git repository history, the set of commits where it was introduced and the accompanying pathname for the blob is collected. This is enabled by default, but can be controlled using the new `--git-blob-provenance={first-seen,minimal}` parameter.

Fixes #16.

There are several other improvements and new features in this commit:

* Add a new rule to detect Amazon ARN strings
* Fix a typo in the Okta API Key rule
* Update dependencies