Fix a provenance / metadata collection bug; improve datastore recording
tl;dr: the `--blob-metadata=matching` option was broken, and the
datastore recording logic was badly written.

Previously, when a blob was seen multiple times during scanning, its
metadata and provenance was recorded, regardless of whether the the blob
had any matches or not. The net result was that much more Git commit
metadata and provenance information was recorded than intended.

This commit fixes that bug by storing an additional bit of information
in the global `seen_blobs` data structure that is maintained while
scanning: now, it is recorded whether or not there were matches found
when scanning a blob, not just whether or not the blob was seen before.

Additionally, this commit refines the datastore recording
implementation, resulting in far less disk writes than before.