Add an improved ruleset selection mechanism.
This commit introduces a new "ruleset" type, which is a named collection
of rules that can be enabled as a group. Two built-in rulesets are
provided: `np.default`, which is enabled by default, and `np.assets`,
which enables a number of rules that detect non-secret assets (such as
cloud buckets).

Previously, Nosey Parker's rules were an all-or-nothing thing: any rule
that Nosey Parker knew about would be enabled during scanning.

Fixes #51.