Fix 3 existing rules
This should eliminate error messages like `Regex failed to match where
vectorscan did; something is probably odd about the rule`.

The underlying problem is some difference in handling of `\b` anchors
between Rust's `regex` crate and Vectorscan, when the input ends with a
`-` character. This is problematic with rules like `[a-f0-9-]{4} \b`,
which should match 4 hexadecimal or hyphen characters, followed by a
word anchor. Vectorscan treats an input like `abc-` as matching, but the
`regex` crate does not.

In such cases, a more complicated anchor-type expression needs to be
used: the negation of the character class OR end-of-input, e.g.,
`(?: [^a-f0-9-] | $ )`.