Adjust schema for older sqlite versions
The support for ORDER BY clauses in aggregates was only added in sqlite
3.44.

To avoid inconveniencing other consumers of Nosey Parker datastores
(such as older Python clients that only have sqlite 3.43), avoid using
the ORDER BY in the aggregate.

In this case, the change may produce visibly different query results.
However, the order of that data is insignificant.